\name{alanineUnknownSpectrum}
\alias{alanineUnknownSpectrum}
\alias{alanineAuthenticSpectrum}
\docType{data}
\title{Example Fragmentation Spectra}
\description{Examples for \code{\link{SpectrumSimilarity}}.}
\usage{
alanineUnknownSpectrum
alanineAuthenticSpectrum
}
\format{A data frame consisting of (centroid) \emph{m/z} values in the first column and peak intensities in the second.}
\source{Mass spectra of alanine derivatized with N-Methyl-N-[tert-butyldimethyl-silyl]trifluoroacetimide (MTBSTFA) collected in electron impact (EI) mode on a 2 dimensional gas chromatography time of flight (GCxGC-TOF) mass spectrometer.  alanineUnknownSpectrum was acquired from a human plasma extract.  alanineAuthenticSpectrum was acquired from an authentic material.}
\author{Nathan G. Dodder, NIST}
\keyword{datasets}
