\name{chrom_example_1}
\alias{chrom_example_1}
%\alias{chrom_example_2}
\docType{data}
\title{Example Chromatograms}
\description{Examples for \code{\link{drawChrom}}.}
\usage{
chrom_example_1
%chrom_example_2
}
\format{A data frame consisting of time (minutes) in the first column and intensity (counts per second) in the second.}
\source{Both examples are of peptides acquired on a nanoflow liquid chromatography system coupled to a triple quadrupole mass spectrometer operated in multiple reaction monitoring mode.}
\author{Nathan G. Dodder, NIST}
\keyword{datasets} 
