\name{chromatogramExample1}
\alias{chromatogramExample1}
\alias{chromatogramExample2}
\docType{data}
\title{Example Chromatograms}
\description{Examples for \code{\link{DrawChromatogram}}.}
\usage{
chromatogramExample1
chromatogramExample2
}
\format{A data frame consisting of time (minutes) in the first column and intensity (counts per second) in the second.}
\author{Nathan G. Dodder, NIST}
\keyword{datasets} 
