\name{NIDALSGMEGR_spec}
\alias{NIDALSGMEGR_spec}
\docType{data}
\title{Example Fragmentation Spectra}
\description{Examples for \code{\link{specSimilarity}}.}
\usage{
NIDALSGMEGR_spec
}
\format{A data frame consisting of (centroid) \emph{m/z} values in the first column and peak intensities in the second.}
\source{Peptide sequence NIDALSGMEGR.  Acquired on a hybrid triple quadruploe/linear ion trap mass spectrometer in collision induced dissociation (CID) mode.}
\keyword{datasets}
