\name{example.chromatogram.single}
\alias{example.chromatogram.single}
\alias{example.chromatogram.multiple}
\docType{data}
\title{Example Chromatograms}
\description{Examples for \code{\link{DrawChromatogram}}.}
\usage{
example.chromatogram.single
example.chromatogram.multiple
}
\format{A data frame consisting of time (minutes) in the first column and intensity (counts per second) in the second.}
\author{Nathan G. Dodder}
\keyword{datasets} 
