\name{example.sequence}
\alias{example.sequence}
\docType{data}
\title{Human Serum Albumin Amino Acid Sequence}
\description{Used in examples.}
\usage{example.sequence}
\format{Character sequence.}
\source{Downloaded from the ExPASy Proteomics Server (P02768) on 2009-02-07.  Residues 25-609.}
\keyword{datasets} 
