\name{example.spectrum.labeled}
\alias{example.spectrum.labeled}
\docType{data}
\title{Example MALDI-TOF Spectrum of Nitrogen-15 Labeled Peptide} 
\description{Example used in Vinegette.}
\usage{example.spectrum.labeled}
\format{A data frame consisting of \emph{m/z} values in the first column and intensity in the second.}
\source{
Precursor ion spectrum of peptide YEVQGEVFTKPQLWP acquired on a MALDI-TOF/TOF.  The peptide is labeled with approximately 99\% nitrogen-15.
}
\author{Protein expressed by Wei-Li Liao.}
\keyword{datasets}
