\name{example.spectrum.peptide}
\alias{example.spectrum.peptide}
\docType{data}
\title{Example Peptide Fragmentation Spectrum}
\description{Example for \code{\link{PeptideSpectrum}}.}
\usage{
example.spectrum.peptide
}
\format{A data frame consisting of (centroid) \emph{m/z} values in the first column and peak intensities in the second.}
\source{Peptide sequence NIDALSGMEGR.  Acquired on a hybrid triple quadruploe/linear ion trap mass spectrometer in collision induced dissociation (CID) mode.}
\keyword{datasets}
