\name{example.spectrum.unknown}
\alias{example.spectrum.unknown}
\alias{example.spectrum.authentic}
\docType{data}
\title{Example EI Fragmentation Spectra}
\description{Examples for \code{\link{SpectrumSimilarity}}.}
\usage{
example.spectrum.unknown
example.spectrum.authentic
}
\format{A data frame consisting of (centroid) \emph{m/z} values in the first column and peak intensities in the second.}
\source{Mass spectra of alanine derivatized with N-Methyl-N-[tert-butyldimethyl-silyl]trifluoroacetimide (MTBSTFA) collected in electron impact (EI) mode on a 2 dimensional gas chromatography time of flight (GCxGC-TOF) mass spectrometer.  example.spectrum.unknown was acquired from a human plasma extract.  example.spectrum.authentic was acquired from an authentic material.}
\author{Spectra acquired by Gauthier Eppe.}
\keyword{datasets}
