\name{mzCalc-package}
\alias{mzCalc-package}
\alias{mzCalc}
\docType{package}
\title{
A Package for Organic/Biological Mass Spectrometry Calculations and Plots
}
\description{
This package was created to work with mass spectrometry data and computations in a customizable way.
}
\details{
\tabular{ll}{
Package: \tab mzCalc\cr
Type: \tab Package\cr
Version: \tab 0.1\cr
Date: \tab 2009-03-01\cr
License: \tab Unlimited\cr
LazyLoad: \tab yes\cr
}
See the vignette for an overview of the package.
}
\author{
Nathan G. Dodder and Katharine M. Mullen, NIST,
Maintainer: Nathan Dodder <nathan.dodder@nist.gov>
}
\keyword{ package }
