\name{OrgMassSpecR-package}
\alias{OrgMassSpecR-package}
\alias{OrgMassSpecR}
\docType{package}
\title{
A Package for Organic/Biological Mass Spectrometry Calculations and Plots
}
\description{
Use the command \code{help(package = "OrgMassSpecR")} to see an overview of the package.
}
\keyword{ package }
