\name{OrgMassSpecR-package}
\alias{OrgMassSpecR-package}
\alias{OrgMassSpecR}
\docType{package}
\title{
Organic Mass Spectrometry
}
\description{
Organic/biological mass spectrometry data analysis.

http://orgmassspecr.r-forge.r-project.org/

Use the command \code{help(package = "OrgMassSpecR")} to see an overview of the package.  Use the command \code{vignette(package = "OrgMassSpecR")} to view the package vignette.
}
\keyword{ package }
