\name{OrgMassSpecR-package}
\alias{OrgMassSpecR-package}
\alias{OrgMassSpecR}
\docType{package}
\title{
Organic/Biological Mass Spectrometry
}
\description{
Use the command \code{help(package = "OrgMassSpecR")} to see an overview of the package.  Use the command \code{vignette(package = "OrgMassSpecR")} to view the package vignette.
}
\keyword{ package }
